# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Clima",
        "description": "Documentación de la API para obtener información del clima, registrar usuarios y marcar ciudades favoritas.",
        "contact": {
            "email": "tucorreo@dominio.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor principal"
        }
    ],
    "paths": {
        "/api/auth/weather": {
            "get": {
                "tags": [
                    "Clima"
                ],
                "summary": "Consultar clima por ciudad",
                "operationId": "b2bda65665e62b85b6cef7df54a79d84",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Nombre de la ciudad",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bogotá"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del clima obtenidos correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "temperature": {
                                            "type": "number",
                                            "example": 20
                                        },
                                        "condition": {
                                            "type": "string",
                                            "example": "Sunny"
                                        },
                                        "wind_kph": {
                                            "type": "number",
                                            "example": 10
                                        },
                                        "humidity": {
                                            "type": "number",
                                            "example": 60
                                        },
                                        "local_time": {
                                            "type": "string",
                                            "example": "2025-04-07 10:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ciudad no encontrada o error externo"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/history": {
            "get": {
                "tags": [
                    "Historial"
                ],
                "summary": "Obtener historial de búsquedas del usuario",
                "operationId": "8cd5ffe8cc07dff04a2aae4d468dd6d4",
                "responses": {
                    "200": {
                        "description": "Lista de búsquedas realizadas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "city": {
                                                "type": "string",
                                                "example": "Bogotá"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "2025-04-07T12:00:00.000000Z"
                                            },
                                            "is_favorite": {
                                                "type": "boolean",
                                                "example": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/favorites/{id}": {
            "post": {
                "tags": [
                    "Favoritos"
                ],
                "summary": "Marcar o desmarcar una búsqueda como favorita",
                "operationId": "930375d8ecbf8cf0680b5a97f21df251",
                "parameters": [
                    {
                        "name": "search_id",
                        "in": "path",
                        "description": "ID de la búsqueda",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ciudad marcada como favorita"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Búsqueda no encontrada"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/favorites": {
            "get": {
                "tags": [
                    "Favoritos"
                ],
                "summary": "Obtener ciudades favoritas del usuario",
                "operationId": "3c711801e9946a69a0ce53b76b39be9c",
                "responses": {
                    "200": {
                        "description": "Lista de búsquedas marcadas como favoritas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "city": {
                                                "type": "string",
                                                "example": "Medellín"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "2025-04-07T14:00:00.000000Z"
                                            },
                                            "is_favorite": {
                                                "type": "boolean",
                                                "example": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar un nuevo usuario",
                "operationId": "da8643013c2941f5ed7f8a9c9b4a7f7a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Juan Pérez"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "juan@example.com"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2025-04-07T12:00:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2025-04-07T12:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|sadasdJHASD...ASDqweqwe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "operationId": "535ffeda4b78916efd658b2844ea897a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Juan Pérez"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "juan@example.com"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2025-04-07T12:00:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2025-04-07T12:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "2|jhjhgASDJASDASDhj"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión del usuario autenticado",
                "operationId": "46f8a14c8083abb69eed72768aa08629",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sesión cerrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Clima",
            "description": "Clima"
        },
        {
            "name": "Historial",
            "description": "Historial"
        },
        {
            "name": "Favoritos",
            "description": "Favoritos"
        },
        {
            "name": "Autenticación",
            "description": "Autenticación"
        }
    ]
}